Full code cleanup with fucntions split to different classes
